+++ app/web/src/main/java/uk/ac/ebi/fg/annotare2/web/gwt/user/client/view/SubmissionListView.java
@@ -31,6 +31,6 @@ public interface SubmissionListView extends IsWidget {
     void setSubmissions(List<SubmissionRow> submissions);
 
     public interface Presenter  {
-        void onSubmissionSelected(int id);
+        void onSubmissionSelected(long id);
     }
 }
